requirements.txt
python-dotenv==0.19.0
apify-client==1.6.3
textblob==0.17.1
openai==1.12.0
requests==2.31.0
crewai==0.11.0
python-multipart==0.0.6
Pillow==10.2.0
nltk==3.8.1
tqdm==4.66.1
aiohttp==3.9.3
asyncio==3.4.3
PyYAML==6.0.1